+++ src/main/java/net/earlystage/init/TagInit.java
@@ -12,6 +12,7 @@ public class TagInit {
     // Item
     public static final TagKey<Item> USABLE_CRAFTING_ROCK_ITEMS = TagKey.of(RegistryKeys.ITEM, Identifier.of("earlystage", "usable_crafting_rock_items"));
     public static final TagKey<Item> BARK_ITEMS = TagKey.of(RegistryKeys.ITEM, Identifier.of("earlystage", "bark_items"));
+    public static final TagKey<Item> ROCK_ITEMS = TagKey.of(RegistryKeys.ITEM, Identifier.of("earlystage", "rock_items"));
     // Block
     public static final TagKey<Block> ROCK_FEATURE_BLOCKS = TagKey.of(RegistryKeys.BLOCK, Identifier.of("earlystage", "rock_feature_blocks"));
     // Biome