+++ src/main/java/net/earlystage/mixin/ShovelItemMixin.java
@@ -22,13 +22,14 @@ public class ShovelItemMixin {
     @Inject(method = "useOnBlock", at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/World;getBlockState(Lnet/minecraft/util/math/BlockPos;)Lnet/minecraft/block/BlockState;", ordinal = 0), cancellable = true, locals = LocalCapture.CAPTURE_FAILSOFT)
     private void useOnBlockMixin(ItemUsageContext context, CallbackInfoReturnable<ActionResult> info, World world, BlockPos blockPos, BlockState blockState) {
         if (blockState.isOf(BlockInit.ROCK) || blockState.isOf(BlockInit.FLINT)) {
-            if (!world.isClient) {
-                if (blockState.isOf(BlockInit.ROCK))
+            if (!world.isClient()) {
+                if (blockState.isOf(BlockInit.ROCK)) {
                     ((RockBlock) blockState.getBlock()).cycleState(blockState, world, blockPos);
-                else
+                } else {
                     ((FlintBlock) blockState.getBlock()).cycleState(blockState, world, blockPos);
+                }
             }
-            info.setReturnValue(ActionResult.success(world.isClient));
+            info.setReturnValue(ActionResult.success(world.isClient()));
         }
     }