+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Added:
-- 
+- Added bark compat:
+  - natures spirit
+  - oblivion
+  - regions unexplored
+  - terrestria
+- Added bigglobe rock compat
 ### Fixed:
 - Fixed brick bucket deleting waterlogged block
 ### Changed: